+++ backend/manager/modules/restapi/jaxrs/src/test/java/org/ovirt/engine/api/restapi/resource/BackendTemplatesResourceTest.java
@@ -383,7 +383,7 @@ public class BackendTemplatesResourceTest
         assertEquals(GUIDS[2].toString(), model.getCluster().getId());
         assertNotNull(model.getCpu());
         assertNotNull(model.getCpu().getTopology());
-        assertEquals(4, model.getCpu().getTopology().getCores());
-        assertEquals(2, model.getCpu().getTopology().getSockets());
+        assertEquals(4, model.getCpu().getTopology().getCores().intValue());
+        assertEquals(2, model.getCpu().getTopology().getSockets().intValue());
     }
 }
